+++ src/constants/account.ts
@@ -40,33 +40,3 @@ export const eventTypeToVoucherType = {
   payment: 'expense' as VoucherType,
   transfer: 'transfer' as VoucherType,
 };
-
-// ToDo: move to util (20240516 - Murky)
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function isEventType(data: any): data is EventType {
-  return Object.values(EventType).includes(data);
-}
-
-// ToDo: move to util (20240516 - Murky)
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function isVoucherType(data: any): data is VoucherType {
-  return Object.values(VoucherType).includes(data);
-}
-
-// ToDo: move to util (20240516 - Murky)
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function isPaymentStatusType(data: any): data is PaymentStatusType {
-  return Object.values(PaymentStatusType).includes(data);
-}
-
-// ToDo: move to util (20240516 - Murky)
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function isPaymentPeriodType(data: any): data is PaymentPeriodType {
-  return Object.values(PaymentPeriodType).includes(data);
-}
-
-// ToDo: move to util (20240516 - Murky)
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function isProgressStatus(data: any): data is ProgressStatus {
-  return Object.values(ProgressStatus).includes(data as ProgressStatus);
-}
