+++ src/interfaces/payment.ts
@@ -1,9 +1,9 @@
 import {
   PaymentPeriodType,
   PaymentStatusType,
-  isPaymentPeriodType,
-  isPaymentStatusType,
 } from '@/constants/account';
+
+import { isPaymentPeriodType, isPaymentStatusType } from '@/lib/utils/type_guard/constants/account';
 
 export interface IPayment {
   isRevenue: boolean; // 是否會創造收入，true是錢會進來，false是錢會出去
