+++ tsconfig.json
@@ -1,7 +1,11 @@
 {
   "compilerOptions": {
     "target": "es5",
-    "lib": ["dom", "dom.iterable", "esnext"],
+    "lib": [
+      "dom",
+      "dom.iterable",
+      "esnext"
+    ],
     "allowJs": true,
     "skipLibCheck": true,
     "strict": true,
@@ -19,7 +23,9 @@
         "src/*",
         "prisma/*",
       ],
-      "@package": ["package.json"],
+      "@package": [
+        "package.json"
+      ],
     }
   },
   "include": [
@@ -29,5 +35,7 @@
     "tailwind.config.ts",
     "src/lib/utils/email.js"
   ],
-  "exclude": ["node_modules"]
+  "exclude": [
+    "node_modules"
+  ]
 }
